requirements.txt
astroid==1.6.5
backports.functools-lru-cache==1.5
click==6.7
colorama==0.3.9
configparser==3.5.0
enum34==1.1.6
futures==3.2.0
glfw==1.6.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.2.0
PyGLM==0.4.5b1
pylint==1.9.2
PyOpenGL==3.1.0
singledispatch==3.4.0.3
six==1.11.0
wrapt==1.10.11
